fixed GiftCertificateDaoImpl and TagDao

+++ module2/dao/src/main/java/com/epam/esm/dao/TagDao.java
@@ -9,5 +9,5 @@ public interface TagDao extends AbstractDao<Tag> {
 
     Optional<Tag> findByName(String name);
 
-    List<Tag> findTagsByGiftCertificateId(Integer giftCertificateId);
+    List<Tag> findTagsByGiftCertificateId(Long giftCertificateId);
 }
